Add Single Server adapter
The Single Server adapter has been available for some time, but never listed anywhere for testing/use. This PR adds it.

Of course, it also does a large amount of reformatting, to improve readability. I found that it was difficult to locate the boundary between adapters, so I adjusted the formatting to make it easier to scan the page and find things quickly.

Also updated the Linode section, as it was pretty outdated by now.